+++ ergo_std/ChangeLog.md
@@ -22,6 +22,7 @@
 * Add `std:fs:archive` for creating archives (opposite of `std:fs:unarchive`).
 * Add `std:json` module with `parse` and `stringify` functions to convert json
   strings to ergo values and visa versa.
+* Add `std:string:chars` to get the characters in a string.
 
 ### Improvements
 * `std:match` now takes an array of cases rather than using the remaining